Simplified sparse graph
