+++ jgrapht-opt/src/main/java/org/jgrapht/opt/graph/sparse/SparseDirectedGraph.java
@@ -17,14 +17,21 @@
  */
 package org.jgrapht.opt.graph.sparse;
 
-import java.io.Serializable;
 import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
+import java.util.Set;
+import java.util.function.Supplier;
 
+import org.jgrapht.GraphType;
 import org.jgrapht.alg.util.Pair;
+import org.jgrapht.graph.AbstractGraph;
+import org.jgrapht.graph.DefaultGraphType;
+import org.jgrapht.util.UnmodifiableUnionSet;
 
 /**
- * Sparse unmodifiable directed graph.
+ * Base implementation of a sparse unmodifiable directed graph.
  *
  * <p>
  * Assuming the graph has $n$ vertices, the vertices are numbered from $0$ to $n-1$. Similarly,
@@ -32,22 +39,20 @@ import org.jgrapht.alg.util.Pair;
  * 
  * <p>
  * It stores two boolean incidence matrix of the graph (rows are vertices and columns are edges) as
- * Compressed Sparse Row (CSR). This is a classic format for write-once read-many use cases. Thus,
- * the graph is unmodifiable. In order to also support constant time source and target lookups from
- * an edge identifier we also store the two transposed incidence matrices again in compressed sparse
- * row format.
+ * Compressed Sparse Row (CSR) together with their transpose matrices. This is a classic format for
+ * write-once read-many use cases. Thus, the graph is unmodifiable.
  * 
  * @author Dimitrios Michail
  */
 public class SparseDirectedGraph
     extends
-    BaseSparseDirectedGraph
-    implements
-    Serializable
+    AbstractGraph<Integer, Integer>
 {
-    private static final long serialVersionUID = 7740349500018476517L;
+    protected static final String UNMODIFIABLE = "this graph is unmodifiable";
 
+    protected CSRBooleanMatrix outIncidenceMatrix;
     protected CSRBooleanMatrix outIncidenceMatrixT;
+    protected CSRBooleanMatrix inIncidenceMatrix;
     protected CSRBooleanMatrix inIncidenceMatrixT;
 
     /**
@@ -58,20 +63,145 @@ public class SparseDirectedGraph
      */
     public SparseDirectedGraph(int numVertices, List<Pair<Integer, Integer>> edges)
     {
-        super(numVertices, edges);
-
+        List<Pair<Integer, Integer>> outgoing = new ArrayList<>();
         List<Pair<Integer, Integer>> outgoingT = new ArrayList<>();
+        List<Pair<Integer, Integer>> incoming = new ArrayList<>();
         List<Pair<Integer, Integer>> incomingT = new ArrayList<>();
 
         int eIndex = 0;
         for (Pair<Integer, Integer> e : edges) {
+            outgoing.add(Pair.of(e.getFirst(), eIndex));
             outgoingT.add(Pair.of(eIndex, e.getFirst()));
+            incoming.add(Pair.of(e.getSecond(), eIndex));
             incomingT.add(Pair.of(eIndex, e.getSecond()));
             eIndex++;
         }
 
+        outIncidenceMatrix = new CSRBooleanMatrix(numVertices, edges.size(), outgoing);
         outIncidenceMatrixT = new CSRBooleanMatrix(edges.size(), numVertices, outgoingT);
+        inIncidenceMatrix = new CSRBooleanMatrix(numVertices, edges.size(), incoming);
         inIncidenceMatrixT = new CSRBooleanMatrix(edges.size(), numVertices, incomingT);
+    }
+
+    @Override
+    public Supplier<Integer> getVertexSupplier()
+    {
+        return null;
+    }
+
+    @Override
+    public Supplier<Integer> getEdgeSupplier()
+    {
+        return null;
+    }
+
+    @Override
+    public Integer addEdge(Integer sourceVertex, Integer targetVertex)
+    {
+        throw new UnsupportedOperationException(UNMODIFIABLE);
+    }
+
+    @Override
+    public boolean addEdge(Integer sourceVertex, Integer targetVertex, Integer e)
+    {
+        throw new UnsupportedOperationException(UNMODIFIABLE);
+    }
+
+    @Override
+    public Integer addVertex()
+    {
+        throw new UnsupportedOperationException(UNMODIFIABLE);
+    }
+
+    @Override
+    public boolean addVertex(Integer v)
+    {
+        throw new UnsupportedOperationException(UNMODIFIABLE);
+    }
+
+    @Override
+    public boolean containsEdge(Integer e)
+    {
+        return e >= 0 && e < outIncidenceMatrix.columns();
+    }
+
+    @Override
+    public boolean containsVertex(Integer v)
+    {
+        return v >= 0 && v < outIncidenceMatrix.rows();
+    }
+
+    @Override
+    public Set<Integer> edgeSet()
+    {
+        return new IntegerSet(outIncidenceMatrix.columns());
+    }
+
+    @Override
+    public int degreeOf(Integer vertex)
+    {
+        assertVertexExist(vertex);
+        return outIncidenceMatrix.nonZeros(vertex) + inIncidenceMatrix.nonZeros(vertex);
+    }
+
+    @Override
+    public Set<Integer> edgesOf(Integer vertex)
+    {
+        assertVertexExist(vertex);
+        return new UnmodifiableUnionSet<>(
+            outIncidenceMatrix.nonZerosSet(vertex), inIncidenceMatrix.nonZerosSet(vertex));
+    }
+
+    @Override
+    public int inDegreeOf(Integer vertex)
+    {
+        assertVertexExist(vertex);
+        return inIncidenceMatrix.nonZeros(vertex);
+    }
+
+    @Override
+    public Set<Integer> incomingEdgesOf(Integer vertex)
+    {
+        assertVertexExist(vertex);
+        return inIncidenceMatrix.nonZerosSet(vertex);
+    }
+
+    @Override
+    public int outDegreeOf(Integer vertex)
+    {
+        assertVertexExist(vertex);
+        return outIncidenceMatrix.nonZeros(vertex);
+    }
+
+    @Override
+    public Set<Integer> outgoingEdgesOf(Integer vertex)
+    {
+        assertVertexExist(vertex);
+        return outIncidenceMatrix.nonZerosSet(vertex);
+    }
+
+    @Override
+    public Integer removeEdge(Integer sourceVertex, Integer targetVertex)
+    {
+        throw new UnsupportedOperationException(UNMODIFIABLE);
+    }
+
+    @Override
+    public boolean removeEdge(Integer e)
+    {
+        throw new UnsupportedOperationException(UNMODIFIABLE);
+    }
+
+    @Override
+    public boolean removeVertex(Integer v)
+    {
+        throw new UnsupportedOperationException(UNMODIFIABLE);
+    }
+
+    @Override
+    public Set<Integer> vertexSet()
+    {
+        return new IntegerSet(outIncidenceMatrix.rows());
     }
 
     @Override
@@ -86,6 +216,101 @@ public class SparseDirectedGraph
     {
         assertEdgeExist(e);
         return inIncidenceMatrixT.nonZerosPositionIterator(e).next();
+    }
+
+    @Override
+    public GraphType getType()
+    {
+        return new DefaultGraphType.Builder()
+            .directed().weighted(false).modifiable(false).allowMultipleEdges(true)
+            .allowSelfLoops(true).build();
+    }
+
+    @Override
+    public double getEdgeWeight(Integer e)
+    {
+        return 1.0;
+    }
+
+    @Override
+    public void setEdgeWeight(Integer e, double weight)
+    {
+        throw new UnsupportedOperationException(UNMODIFIABLE);
+    }
+
+    @Override
+    public Integer getEdge(Integer sourceVertex, Integer targetVertex)
+    {
+        if (sourceVertex < 0 || sourceVertex >= outIncidenceMatrix.rows()) {
+            return null;
+        }
+        if (targetVertex < 0 || targetVertex >= outIncidenceMatrix.rows()) {
+            return null;
+        }
+
+        Iterator<Integer> it = outIncidenceMatrix.nonZerosPositionIterator(sourceVertex);
+        while (it.hasNext()) {
+            int eId = it.next();
+            if (getEdgeTarget(eId) == targetVertex) {
+                return eId;
+            }
+        }
+        return null;
+    }
+
+    @Override
+    public Set<Integer> getAllEdges(Integer sourceVertex, Integer targetVertex)
+    {
+        if (sourceVertex < 0 || sourceVertex >= outIncidenceMatrix.rows()) {
+            return null;
+        }
+        if (targetVertex < 0 || targetVertex >= outIncidenceMatrix.rows()) {
+            return null;
+        }
+
+        Set<Integer> result = new HashSet<>();
+
+        Iterator<Integer> it = outIncidenceMatrix.nonZerosPositionIterator(sourceVertex);
+        while (it.hasNext()) {
+            int eId = it.next();
+
+            if (getEdgeTarget(eId) == targetVertex) {
+                result.add(eId);
+            }
+        }
+        return result;
+    }
+
+    /**
+     * Ensures that the specified vertex exists in this graph, or else throws exception.
+     *
+     * @param v vertex
+     * @return <code>true</code> if this assertion holds.
+     * @throws IllegalArgumentException if specified vertex does not exist in this graph.
+     */
+    protected boolean assertVertexExist(Integer v)
+    {
+        if (v >= 0 && v < outIncidenceMatrix.rows()) {
+            return true;
+        } else {
+            throw new IllegalArgumentException("no such vertex in graph: " + v.toString());
+        }
+    }
+
+    /**
+     * Ensures that the specified edge exists in this graph, or else throws exception.
+     *
+     * @param e edge
+     * @return <code>true</code> if this assertion holds.
+     * @throws IllegalArgumentException if specified edge does not exist in this graph.
+     */
+    protected boolean assertEdgeExist(Integer e)
+    {
+        if (e >= 0 && e < outIncidenceMatrix.columns()) {
+            return true;
+        } else {
+            throw new IllegalArgumentException("no such edge in graph: " + e.toString());
+        }
     }
 
 }
