+++ jgrapht-opt/src/main/java/org/jgrapht/opt/graph/sparse/SparseDirectedWeightedGraph.java
@@ -18,8 +18,6 @@
 package org.jgrapht.opt.graph.sparse;
 
 import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -44,20 +42,22 @@ import org.jgrapht.alg.util.Triple;
  * Compressed Sparse Row (CSR). This is a classic format for write-once read-many use cases. Thus,
  * the graph is unmodifiable. In order to also support constant time source and target lookups from
  * an edge identifier we also store the two transposed incidence matrices again in compressed sparse
- * row format. In these transposed matrices we also maintain the edge weights.
+ * row format. The edge weights are maintained in an array indexed by the edge identifier.
  * 
  * @author Dimitrios Michail
  */
 public class SparseDirectedWeightedGraph
     extends
-    BaseSparseDirectedGraph
+    SparseDirectedGraph
     implements
     Serializable
 {
     private static final long serialVersionUID = -7601401110000642281L;
 
-    protected CSRDoubleMatrix outIncidenceMatrixT;
-    protected CSRDoubleMatrix inIncidenceMatrixT;
+    /**
+     * The edge weights
+     */
+    protected double[] weights;
 
     /**
      * Create a new graph from an edge list.
@@ -74,33 +74,13 @@ public class SparseDirectedWeightedGraph
                 .stream().map(e -> Pair.of(e.getFirst(), e.getSecond()))
                 .collect(Collectors.toList()));
 
-        List<Triple<Integer, Integer, Double>> outgoingT = new ArrayList<>();
-        List<Triple<Integer, Integer, Double>> incomingT = new ArrayList<>();
+        this.weights = new double[edges.size()];
 
         int eIndex = 0;
         for (Triple<Integer, Integer, Double> e : edges) {
             double edgeWeight = e.getThird() != null ? e.getThird() : Graph.DEFAULT_EDGE_WEIGHT;
-            outgoingT.add(Triple.of(eIndex, e.getFirst(), edgeWeight));
-            incomingT.add(Triple.of(eIndex, e.getSecond(), edgeWeight));
-            eIndex++;
+            weights[eIndex++] = edgeWeight;
         }
-
-        outIncidenceMatrixT = new CSRDoubleMatrix(edges.size(), numVertices, outgoingT);
-        inIncidenceMatrixT = new CSRDoubleMatrix(edges.size(), numVertices, incomingT);
-    }
-
-    @Override
-    public Integer getEdgeSource(Integer e)
-    {
-        assertEdgeExist(e);
-        return outIncidenceMatrixT.nonZerosPositionIterator(e).next();
-    }
-
-    @Override
-    public Integer getEdgeTarget(Integer e)
-    {
-        assertEdgeExist(e);
-        return inIncidenceMatrixT.nonZerosPositionIterator(e).next();
     }
 
     @Override
@@ -113,21 +93,14 @@ public class SparseDirectedWeightedGraph
     public double getEdgeWeight(Integer e)
     {
         assertEdgeExist(e);
-
-        Iterator<Pair<Integer, Double>> it = outIncidenceMatrixT.nonZerosIterator(e);
-        if (it.hasNext()) {
-            return it.next().getSecond();
-        }
-        return inIncidenceMatrixT.nonZerosIterator(e).next().getSecond();
+        return weights[e];
     }
 
     @Override
     public void setEdgeWeight(Integer e, double weight)
     {
         assertEdgeExist(e);
-
-        outIncidenceMatrixT.setNonZeros(e, weight);
-        inIncidenceMatrixT.setNonZeros(e, weight);
+        weights[e] = weight;
     }
 
 }
